+++ CHANGELOG.md
@@ -24,7 +24,8 @@ N/A
 
 ### Changed
 
-Updated data structure for static-content section in rest.yaml
+1. Updated data structure for static-content section in rest.yaml
+2. Fixed bug for setting multiple HTTP cookies
 
 ---
 ## Version 3.0.14, 4/28/2024